remove the sync client
